# src/api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Multi-Chain Smart Contract Analytics API
  description: |
    Comprehensive blockchain analytics platform that analyzes smart contracts across multiple chains 
    with intelligent chain isolation and enterprise-grade reliability.
    
    ## Features
    - **Multi-Chain Support**: Ethereum, Lisk, Starknet, Polygon, Base, Arbitrum, Optimism
    - **OAuth Authentication**: Google and GitHub OAuth integration
    - **Multiple Contract Configurations**: Save and manage multiple analysis configurations
    - **Analysis History**: Track all your analyses with detailed results
    - **Comprehensive Analytics**: 20+ DeFi metrics across user lifecycle, activity, financial, and performance categories
    - **Multiple Output Formats**: JSON, CSV, and Markdown reports
    
    ## Authentication
    This API supports two authentication methods:
    1. **JWT Bearer Token**: Use the `/api/auth/login` endpoint to get a token
    2. **API Key**: Use the `X-API-Key` header with your personal API key
    
    ## Rate Limits
    - General endpoints: 100 requests per 15 minutes
    - Analysis endpoints: 10 requests per hour
    - Limits vary by user tier (Free, Pro, Enterprise)
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@analytics-api.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.analytics-platform.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        avatar:
          type: string
          nullable: true
          example: "https://example.com/avatar.jpg"
        tier:
          type: string
          enum: [free, pro, enterprise]
          example: "free"
        oauthProvider:
          type: string
          enum: [google, github]
          nullable: true
          example: "google"
        isActive:
          type: boolean
          example: true
        lastLogin:
          type: string
          format: date-time
          nullable: true
        preferences:
          type: object
          properties:
            notifications:
              type: object
              properties:
                email:
                  type: boolean
                analysis:
                  type: boolean
            defaultChain:
              type: string
              example: "ethereum"
        usage:
          type: object
          properties:
            analysisCount:
              type: integer
              example: 5
            monthlyAnalysisCount:
              type: integer
              example: 3
            lastAnalysis:
              type: string
              format: date-time
              nullable: true
    Contract:
      type: object
      required:
        - address
        - chain
        - name
      properties:
        address:
          type: string
          example: "0x1234567890123456789012345678901234567890"
        chain:
          type: string
          enum: [ethereum, lisk, starknet, polygon, base, arbitrum, optimism]
          example: "ethereum"
        name:
          type: string
          example: "USDT Token"
        abi:
          type: string
          nullable: true
          example: "./abis/usdt.json"
...
    AnalysisResult:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        userId:
          type: string
          example: "507f1f77bcf86cd799439011"
        configId:
          type: string
          example: "507f1f77bcf86cd799439011"
        analysisType:
          type: string
          enum: [single, competitive, comparative]
          example: "single"
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
          example: "completed"
        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 100
        results:
          type: object
          properties:
            target:
              type: object
              properties:
                contract:
                  $ref: '#/components/schemas/Contract'
                transactions:
                  type: integer
                  example: 1250
                metrics:
                  type: object
                  description: "Comprehensive DeFi metrics"
                behavior:
                  type: object
                  description: "User behavior analysis"
                reportPaths:
                  type: object
                  properties:
                    json:
                      type: string
                    csv:
                      type: string
                    markdown:
                      type: string
            competitors:
              type: array
              items:
                type: object
                properties:
                  contract:
                    $ref: '#/components/schemas/Contract'
                  transactions:
                    type: integer
                  metrics:
                    type: object
                  behavior:
                    type: object
                  error:
                    type: string
                    nullable: true
            comparative:
              type: object
              properties:
                summary:
                  type: string
                rankings:
                  type: object
                insights:
                  type: array
                  items:
                    type: string
        metadata:
          type: object
          properties:
            blockRange:
              type: integer
            chainsAnalyzed:
              type: array
              items:
                type: string
            totalTransactions:
              type: integer
            executionTimeMs:
              type: integer
            rpcCallsCount:
              type: integer
            errorCount:
              type: integer
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
...
paths:
...
  /api/users/dashboard:
    get:
      summary: Get user dashboard data
      description: Retrieve comprehensive dashboard data for the user
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  stats:
                    type: object
                    properties:
                      contractConfigs:
                        type: integer
                      totalAnalyses:
                        type: integer
                      completedAnalyses:
                        type: integer
                      monthlyAnalyses:
                        type: integer
                      avgExecutionTime:
                        type: number
                  recentAnalyses:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnalysisResult'
                  usage:
                    type: object
                  limits:
                    type: object